+++ moabb/datasets/utils.py
@@ -149,4 +149,4 @@ def _download_all(update_path=True, verbose=None):
     # iterate over dataset
     for ds in dataset_list:
         # call download
-        ds().download(update_path=True, verbose=verbose)
+        ds().download(update_path=True, verbose=verbose, accept=True)